+++ PlantVsZombies/Game/Board.h
@@ -96,9 +96,9 @@ public:
 	int mCurrentWave = 0;			// 当前波
 	int mMaxWave = 10;		// 关卡总波数
 	float mZombieCountDown = 20.0f;		// 下一波僵尸倒计时
-	long long mTotalZombieHP = 0;		// 在场全部僵尸血量
-	long long mCurrectWaveZombieHP = 0;	// 本波僵尸血量
-	long long mNextWaveSpawnZombieHP = 0;		// 下一波僵尸刷新血量
+	int64_t mTotalZombieHP = 0;		// 在场全部僵尸血量
+	int64_t mCurrectWaveZombieHP = 0;	// 本波僵尸血量
+	int64_t mNextWaveSpawnZombieHP = 0;		// 下一波僵尸刷新血量
 
 	int mZombieNumber = 0;
 	bool mTrophySpawned = false;  // 防止重复生成
