+++ PlantVsZombies/Game/Zombie/BucketZombie.cpp
@@ -22,7 +22,7 @@ void BucketZombie::HelmDrop()
 void BucketZombie::CheckHelmImage()
 {
 	if (mHelmType == HelmType::HELMTYPE_NONE) return;
-	if (mHelmStage == ArmorBrokenState::NO_BROKEN && mHelmHealth <= static_cast<long long>(mHelmMaxHealth) * 2 / 3) {
+	if (mHelmStage == ArmorBrokenState::NO_BROKEN && mHelmHealth <= static_cast<int64_t>(mHelmMaxHealth) * 2 / 3) {
 		mHelmStage = ArmorBrokenState::A_LITTLE_BROKEN;
 		mAnimator->SetTrackImage("anim_bucket", ResourceManager::GetInstance().
 			GetTexture("IMAGE_ZOMBIE_BUCKET2"));
